small bug fixes: query_exec_time in ns and correct query text

+++ receiver/postgresqlreceiver/scraper.go
@@ -192,7 +192,7 @@ func (p *postgreSQLScraper) collectQueryPerfStats(
 
 	for _, s := range queryStats {
 		p.mb.RecordPostgresqlQueryCountDataPoint(now, s.queryCount, s.queryText, s.queryId)
-		p.mb.RecordPostgresqlQueryTotalExecTimeDataPoint(now, s.queryExecTime, s.queryText, s.queryId)
+		p.mb.RecordPostgresqlQueryTotalExecTimeDataPoint(now, int64(s.queryExecTime), s.queryText, s.queryId)
 	}
 }
